appiumworkspace/requirements.txt
Appium-Python-Client==0.26
attrs==17.4.0
more-itertools==4.1.0
pluggy==0.6.0
py==1.5.3
pytest==3.5.0
selenium==3.11.0
six==1.11.0
